+++ arrayjit/lib/utils.ml
@@ -274,7 +274,7 @@ let get_debug name =
         Minidebug_runtime.forget_printbox
         @@ Minidebug_runtime.debug ~time_tagged ~elapsed_times ~location_format ~print_entry_ids
              ~verbose_entry_ids ~global_prefix:name ~toc_entry ~toc_specific_hyperlink:""
-             ~highlight_terms:Re.(alt [ str "wait"; str "release" ])
+             ~highlight_terms:Re.(alt [ str "await" ])
              ~exclude_on_path:Re.(str "env")
              ~log_level ?snapshot_every_sec ()
     | Some filename ->
@@ -283,7 +283,7 @@ let get_debug name =
              ~print_entry_ids ~verbose_entry_ids ~global_prefix:name ~toc_flame_graph:true
              ~flame_graph_separation:50 ~toc_entry ~for_append:false ~max_inline_sexp_length:120
              ~hyperlink ~toc_specific_hyperlink:""
-             ~highlight_terms:Re.(alt [ str "wait"; str "release" ])
+             ~highlight_terms:Re.(alt [ str "await" ])
              ~exclude_on_path:Re.(str "env")
              ~backend ~log_level ?snapshot_every_sec filename