+++ CHANGES.md
@@ -8,6 +8,7 @@
   - We output half precision specific code from the CUDA backend.
 - Finally proper support for mixed precision! Lazy precision defaults and delayed precision setting via `Tnode.update_prec`.
 - A placeholder `nn_blocks.ml` hinting at an intended design pattern for model components.
+- A memory model for the multiple virtual devices per physical device setup, implemented in the CUDA backend. It fixes the CUDA backend behavior in the data parallelism benchmark.
 
 ### Changed